refactor(store): split opensearch.rs into an opensearch/ module directory
